refactor: move function to proper module
