+++ src/files/archives.rs
@@ -1,8 +1,7 @@
-use crate::files::magic::{
-    BZIP2_MAGIC, GZIP_MAGIC, SEVENZ_MAGIC, TAR_MAGIC, TAR_MAGIC_OFFSET, XZ_MAGIC, ZIP_MAGIC,
-};
+use crate::files::magic::*;
 use crate::files::utils::get_file_extension;
 use crate::models::binary_container::BinaryContainer;
+
 use bzip2::read::BzDecoder;
 use flate2::read::GzDecoder;
 use log::{debug, error};
